refactor: import ModalContent
