+++ src/feature/auth/ui/sent-email-modal/index.tsx
@@ -4,6 +4,7 @@ import { useTranslation } from '@/shared/lib'
 import { ReturnComponent } from '@/shared/types'
 import { Button, Modal, Text } from '@/shared/ui'
 import { FlexCol } from '@/shared/ui/flex'
+import { ModalContent } from '@/shared/ui/modal'
 
 type Props = {
   email?: string
@@ -25,7 +26,7 @@ export const SentEmailModal = ({ email, onOpenChange, open }: Props): ReturnComp
 
   return (
     <Modal onOpenChange={onOpenChange} open={open}>
-      <Modal.Content
+      <ModalContent
         asChild
         classNameContent={'max-w-[378px] w-[90vw]'}
         title={t.pages.signUp.modalTitle}
@@ -38,7 +39,7 @@ export const SentEmailModal = ({ email, onOpenChange, open }: Props): ReturnComp
             {t.pages.signUp.modalBtn}
           </Button>
         </FlexCol>
-      </Modal.Content>
+      </ModalContent>
     </Modal>
   )
 }
