+++ src/widgets/add-profile-photo/add-avatar-button/hooks/useUpdateAvatar.ts
